+++ CHANGELOG.md
@@ -26,6 +26,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Account::get_participation_event_ids` method;
 - `RequiredStorageDeposit::{alias(), basic(), foundry(), nft()}` getters;
 - `TransactionOptionsDto`;
+- `Transaction::inputs` and `TransactionDto::inputs` fields;
+- Derive `Eq, PartialEq` for `Account` and `OutputData`;
 
 ### Changed
 
@@ -36,10 +38,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Limit max parallel requests for incoming transactions;
 - Move all participation methods from the AccountManager to the Account;
 - `Account::get_participation_overview` sends requests now in parallel;
+- `Account::{get_incoming_transaction_data(), incoming_transactions()}` return now `Transaction` instead of `(TransactionPayload, Vec<OutputWithMetadataResponse>)`;
+- `AccountDto::incoming_transactions` from `(TransactionPayloadDto, Vec<OutputWithMetadataResponse>)` to `TransactionDto`;
+- `Response::{IncomingTransactionData, IncomingTransactionsData}` contain `TransactionDto` instead of `IncomingTransactionDataDto`;
 
 ### Removed
 
 - `clear_listeners` from the `WalletMessageHandler`;
+- `IncomingTransactionDataDto` type;
 
 ## 1.0.0-rc.4 - 2022-12-23